Changes regarding review
* Autoformatting
* Javadocs
* Renaming
