+++ de.tu_bs.cs.isf.e4cf.compare.data_structures/src/de/tu_bs/cs/isf/e4cf/compare/data_structures/io/reader/JavaReader.java
@@ -15,6 +15,7 @@ import com.github.javaparser.*;
 import com.github.javaparser.ast.*;
 
 /***
+ * This reader converts java files into the generic data structure.
  * 
  * @author Serkan Acar
  * @author Hassan Smaoui
@@ -22,7 +23,6 @@ import com.github.javaparser.ast.*;
  * @author Paulo Haas
  *
  */
-
 public class JavaReader extends AbstractArtifactReader {
 	public final static String[] SUPPORTED_FILE_ENDINGS = { "java" };
 
@@ -30,6 +30,12 @@ public class JavaReader extends AbstractArtifactReader {
 		super(SUPPORTED_FILE_ENDINGS);
 	}
 
+	/**
+	 * Converts java files into a tree.
+	 * 
+	 * @param element Java file
+	 * @return Tree
+	 */
 	@Override
 	public Tree readArtifact(FileTreeElement element) {
 		Tree tree = null;
@@ -38,13 +44,9 @@ public class JavaReader extends AbstractArtifactReader {
 			String s = FileStreamUtil.readLineByLine(Paths.get(element.getAbsolutePath()));
 			String fileName = Paths.get(element.getAbsolutePath()).getFileName().toString();
 			CompilationUnit cu = StaticJavaParser.parse(s);
-
 			Node rootNode = new NodeImpl(fileName);
-			
-			Visitor visitor = new Visitor();
-			
+			JavaVisitor visitor = new JavaVisitor();
 			visitor.visit(cu, rootNode);
-			
 			tree = new TreeImpl(fileName, rootNode);
 
 			// Remove these lines after debug
@@ -56,7 +58,6 @@ public class JavaReader extends AbstractArtifactReader {
 			System.out.print("\n\n--- FRAMEWORK TREE END ---\n\n");
 		}
 
-		
 		return tree;
 	}
 }
